requirements.txt
blinker==1.5
click==8.1.3
Flask==2.2.3
Flask-Mail==0.9.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
Werkzeug==2.2.3
zipp==3.15.0
